chore(.gitignore): add .gitignore file to the project
This commit adds a .gitignore file to the project. The .gitignore file is used to specify intentionally untracked files that Git should ignore. It helps to prevent certain files from being committed to the repository, such as build artifacts, temporary files, and IDE-specific files.

The .gitignore file in this commit was generated using the Toptal Gitignore API for Visual Studio Code and Rust. It includes patterns to ignore files and directories commonly associated with Rust projects, such as debug/ and target/, as well as files generated by build systems like Cargo and rust-analyzer.

Additionally, the .gitignore file includes patterns to ignore Visual Studio Code-specific files and directories, such as .vscode/ and .history/. These files and directories are not relevant to the project and should not be committed to the repository.

For more information on how to use and customize the .gitignore file, please refer to the official Git documentation.